Rework the SCPIPropertyMinMax concept once again.
It was difficult to know whether or not an attribute had min/max
methods, or was a plain SCPIProperty.

There was also the problem that accessing the attribute without querying
"value" directly could result in that the underlying value on the VNA
changed. This could potentially lead to hard-to-find bugs.

The new concept is that SCPIProperty attributes can have an accompanying
attribute with "_minmax" appended to the name. This attribute is a property
with the expected methods, query_min(), etc.